requirements.txt
streamlit
sentence-transformers
faiss-cpu
PyPDF2
pdfplumber
numpy
langchain
google-generativeai
python-dotenv
